Dynamic data updation

+++ src/app/admin/categoryData/categoryData.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { AdminProductsService } from '../adminProducts.service';
 import { HttpClient } from '@angular/common/http';
+import { Title } from '@angular/platform-browser';
 
 @Component({
   selector: 'app-categoryData',
@@ -9,21 +10,21 @@ import { HttpClient } from '@angular/common/http';
 })
 export class CategoryDataComponent implements OnInit {
   categoryDisplay: any = "";
-  categoryCount: any = 0;
-  constructor(private adminService: AdminProductsService, private http: HttpClient) {
+  categoryCount: number = 0;
+  constructor(private adminService: AdminProductsService, private http: HttpClient, private title:Title) {
     this.adminService.getCategory().subscribe((data) => {
       this.categoryDisplay = data;
     });
-    this.adminService.categoryTypesCount().subscribe((category: any) => this.categoryCount = category);
+    this.adminService.categoryTypesCount().subscribe((category: any) => this.categoryCount = category.length);
   }
 
-  ngOnInit(): void { }
+  ngOnInit(): void {
+    this.title.setTitle('Category | RK MART');
+   }
 
   removeCategory(categoryId: any, categoryValue: any, categoryClass: any, categoryUniqueValue: any) {
-    this.adminService.removeCategoryData(categoryId, categoryValue, categoryClass, categoryUniqueValue);
-    // this.adminService.getCategory().subscribe((data) => {
-    //   this.categoryDisplay = data;
-    // });
-    // this.adminService.categoryTypesCount().subscribe((category: any) => this.categoryCount = category);
+    if(confirm(`Are you sure want to delete ${categoryValue}`)){
+      this.adminService.removeCategoryData(categoryId, categoryValue, categoryClass, categoryUniqueValue);
+    }
   }
 }
